- hooked low level buffers into render state.
It still needs to support the old interface so the code isn't really clean

+++ src/gl/system/glsys_vertexbuffer.cpp
@@ -122,7 +122,7 @@ void GLVertexBuffer::SetFormat(int numBindingPoints, int numAttributes, size_t s
 	}
 }
 
-void GLVertexBuffer::Bind(size_t *offsets)
+void GLVertexBuffer::Bind(int *offsets)
 {
 	int i = 0;
