+++ src/gl/system/glsys_vertexbuffer.h
@@ -24,7 +24,7 @@ public:
 	~GLVertexBuffer();
 	void SetData(size_t size, void *data, bool staticdata) override;
 	void SetFormat(int numBindingPoints, int numAttributes, size_t stride, FVertexBufferAttribute *attrs) override;
-	void Bind(size_t *offsets);
+	void Bind(int *offsets);
 	void Map() override;
 	void Unmap() override;
 };
